Add functions to extract info from arguments list
This CLI will be passed a list of arguments, we need to be able to pull out
the command to run and the environment in which to run the command on.

This change set adds two functions which take the argument list and extract
the command and environment from the list.