Add function to generate docker-compose command
This change set adds a function which takes a docker-compose file, validates
it exists then builds the docker-compose command to execute.

Basically it adds the 'docker-compose -f' to the passed in file name.

However, this function will also read the compose file and determin if
it is a v2 or v3, if < v3 it just uses the single -f flag as it assumes
your using the extends stanza; otherwise if the passed in compose file
is v3+ then it will use two -f flags adding both the passed in compose file
(assuming something like docker-compose.test.yml) and the _base_ compose
file (i.e. docker-compose.yml).