Add Run command
This change set adds a Run command which executes the docker-compose run
command with the selected environment compose file but also will ensure that
the --rm flag is always used. This is to avoid being left with loads and loads
of orphan containers which need to be manually pruned.

With the bam run command you can execute tasks against an application via a
new container (think npm install when using volumns) without leaving behind
the transeant container.