Add Lint command
This change set adds the Lint command to bam. This command will execute
a lint test check againts the container.

It will bring up a new container to run the test in then remove the container
once completed. It will always run under _test_ mode and expects that the
container has an NPM command called lint for it to call in order to
execute the linter checks. It is the NPM command (npm run lint) which actually
executes the linter checks.