Add Logs command
This command fetches the logs from the selected container. It is just calling
the docker-compose logs command but ensures that the --follow flag is being
used. It also reads from the config to apply a defualt --tail value if one
is provided.

It also adds a --clear flag which will clear the logs from the container.

This change set also includes an alias for the logs command called log as
the git command to view the commit log is 'log' and I was getting confused
switching between git log and docker-compose logs