Add RM command
This change set adds the RM (remove) command to bam.

It runs the docker-compose rm command but also removes any named or unnamed
volumns used by the container.

The idea here is that if you remove the container the node_moudles folder
which is stored in a named volume is also removed.