Add init command
This change set adds an init command to Bam which can be used to generate or
otherwise _initialize_ a new project.

The init command prompts for input so it knows what kind of project to
initialize. Currently it can just initialize a Nodejs project but plans are
to add more types of project templates.