Add front end question to init command
This chang set updates the init command to ask if the project will include
a front end.

That is will this repo include both the backend and front end, if so then
the genreated project will put the backend under the _backend_ folder and
create a _frontend_ folder for the front end to go.

Note: Currently it does not generate a frontend template