Update template to use bunyan req/res serializers
This change set updates the project template to ensure that when creating
a service project that the logger configuration tells buynan to use the
req/res serializers that it comes with.

These serializers will parse a req/res object down to only a few properties
vs. logging the entire object. This will ensure that the logs are a lot
easer to read as they won't be overloaded with junk data