:memo:Fix OAS auth description in project template
This change set updates the service project template so that the generated
OAS configuration file sets the authenitcation schema description for JWTs
to indicate that we do _not_ need to provide the prefix Bearer when inputting
the auth token. The Swagger UI will prepend the token with Bearer for us.