Update template to include mongo-sanitize
When the service template will include Mongo for the storage provider the
controllers should sanitize the inputs. There is a package for sanitizing
inputs to mongo called mongo-sanitize. This change set just adds the
package to the package.json so that it is available if Mongo is selected
when generating a new service project