Update template to fix bench NPM command
The wild card characters in the NPM bench command were being evaulated to early
and thus was not working correclty.

This change set just updates the template to generate a package.json with
a bench command that does not process the wild cards within the
script/bench script but rather within the NPM command itself.