Update project template to disable promise/always-return
This change set updates the project template to disable the linter rule
promise/always-return which states that all then blocks must return.

This is only done in the TEST code as I often have then blocks which just
run assertions or stores the value returned by the promis in a closure
variable to be used within an after block.