Update the User model to have pending state
This change set updates the user model to include a state property so we can
disable or approve users. This change set also adds a flag to indicate if the
user is a site admin.

This change includes a middleware to be used by routes only accessable by
admin users to ensure that the reuqesting user is an admin. It is exported
by the users module under the name adminGuard