Update project template to check for missing parm on auth
This change set updates the project template to have the passport config
check for requires parameters. This is other then the parameters needed
for passport, that is business logic parameters such as displayName.

If missing a 400 is returned.

It also updates the login passport stategy to only allow _approved_ users
to log in.